backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
email-validator
python-multipart
python-dotenv
firebase-admin
requests
openai
PyPDF2
certifi